+++ pysat/tests/test_orbits.py
@@ -1,5 +1,5 @@
 import datetime as dt
-from dateutil.relativedelta import relativedelta as relativedelta
+from dateutil.relativedelta import relativedelta
 import numpy as np
 
 import pandas as pds
@@ -9,165 +9,138 @@ import pysat
 
 
 class TestOrbitsUserInterface():
+    def setup(self):
+        """ Set up User Interface unit tests
+        """
+        self.in_args = ['pysat', 'testing']
+        self.in_kwargs = {'clean_level': 'clean', 'update_files': True}
+        self.testInst = None
+        self.stime = dt.datetime(2009, 1, 1)
+
+    def teardown(self):
+        """ Tear down user interface tests
+        """
+        del self.in_args, self.in_kwargs, self.testInst, self.stime
 
     def test_orbit_w_bad_kind(self):
-        info = {'index': 'mlt', 'kind': 'cats'}
+        """ Test orbit failure with bad 'kind' input
+        """
+        self.in_kwargs['orbit_info'] = {'index': 'mlt', 'kind': 'cats'}
         with pytest.raises(ValueError):
-            self.testInst = pysat.Instrument('pysat', 'testing',
-                                             clean_level='clean',
-                                             orbit_info=info,
-                                             update_files=True)
+            self.testInst = pysat.Instrument(*self.in_args, **self.in_kwargs)
 
-    def test_orbit_long_w_bad_variable(self):
-        info = {'index': 'magnetic local time', 'kind': 'longitude'}
-        self.testInst = pysat.Instrument('pysat', 'testing',
-                                         clean_level='clean',
-                                         orbit_info=info, update_files=True)
-        self.testInst.load(2009, 1)
+    @pytest.mark.parametrize("info", [({'index': 'magnetic local time',
+                                        'kind': 'longitude'}),
+                                      (None),
+                                      ({'index': 'magnetic local time',
+                                        'kind': 'lt'}),
+                                      ({'index': 'magnetic local time',
+                                       'kind': 'polar'}),
+                                      ({'index': 'magnetic local time',
+                                        'kind': 'orbit'})])
+    def test_orbit_w_bad_orbit_info(self, info):
+        """ Test orbit failure on iteration with orbit initialization
+        """
+        self.in_kwargs['orbit_info'] = info
+        self.testInst = pysat.Instrument(*self.in_args, **self.in_kwargs)
+        self.testInst.load(date=self.stime)
+
         with pytest.raises(ValueError):
             self.testInst.orbits.next()
 
-    def test_orbit_long_w_missing_orbit_info(self):
-        self.testInst = pysat.Instrument('pysat', 'testing',
-                                         clean_level='clean',
-                                         update_files=True)
-        self.testInst.load(2009, 1)
-        with pytest.raises(ValueError):
-            self.testInst.orbits.next()
+    @pytest.mark.parametrize("info", [({'index': 'magnetic local time',
+                                        'kind': 'polar'}),
+                                      ({'index': 'magnetic local time',
+                                        'kind': 'orbit'}),
+                                      ({'index': 'magnetic local time',
+                                        'kind': 'longitude'}),
+                                      ({'index': 'magnetic local time',
+                                        'kind': 'lt'})])
+    def test_orbit_polar_w_missing_orbit_index(self, info):
+        """ Test orbit failure oon iteration with missing orbit index
+        """
+        self.in_kwargs['orbit_info'] = info
+        self.testInst = pysat.Instrument(*self.in_args, **self.in_kwargs)
 
-    def test_orbit_long_w_missing_orbit_index(self):
-        info = {'index': 'magnetic local time', 'kind': 'longitude'}
-        self.testInst = pysat.Instrument('pysat', 'testing',
-                                         clean_level='clean',
-                                         orbit_info=info, update_files=True)
-        # force index to None
+        # Force index to None beforee loading and iterating
         self.testInst.orbits.orbit_index = None
-        self.testInst.load(2009, 1)
+        self.testInst.load(date=self.stime)
         with pytest.raises(ValueError):
             self.testInst.orbits.next()
 
-    def test_orbit_mlt_w_bad_variable(self):
-        info = {'index': 'magnetic local time', 'kind': 'lt'}
-        self.testInst = pysat.Instrument('pysat', 'testing',
-                                         clean_level='clean',
-                                         orbit_info=info, update_files=True)
-        self.testInst.load(2009, 1)
-        with pytest.raises(ValueError):
-            self.testInst.orbits.next()
+    def test_orbit_repr(self):
+        """ Test the Orbit representation
+        """
+        self.in_kwargs['orbit_info'] = {'index': 'mlt'}
+        self.testInst = pysat.Instrument(*self.in_args, **self.in_kwargs)
+        out_str = self.testInst.orbits.__repr__()
 
-    def test_orbit_mlt_w_missing_orbit_index(self):
-        info = {'index': 'magnetic local time', 'kind': 'lt'}
-        self.testInst = pysat.Instrument('pysat', 'testing',
-                                         clean_level='clean',
-                                         orbit_info=info, update_files=True)
-        # force index to None
-        self.testInst.orbits.orbit_index = None
-        self.testInst.load(2009, 1)
-        with pytest.raises(ValueError):
-            self.testInst.orbits.next()
+        assert out_str.find("Orbits(") >= 0
 
-    def test_orbit_polar_w_bad_variable(self):
-        info = {'index': 'magnetic local time', 'kind': 'polar'}
-        self.testInst = pysat.Instrument('pysat', 'testing',
-                                         clean_level='clean',
-                                         orbit_info=info, update_files=True)
-        self.testInst.load(2009, 1)
-        with pytest.raises(ValueError):
-            self.testInst.orbits.next()
+    def test_orbit_str(self):
+        """ Test the Orbit string representation with data
+        """
+        self.in_kwargs['orbit_info'] = {'index': 'mlt'}
+        self.testInst = pysat.Instrument(*self.in_args, **self.in_kwargs)
+        self.testInst.load(date=self.stime)
+        out_str = self.testInst.orbits.__str__()
 
-    def test_orbit_polar_w_missing_orbit_index(self):
-        info = {'index': 'magnetic local time', 'kind': 'polar'}
-        self.testInst = pysat.Instrument('pysat', 'testing',
-                                         clean_level='clean',
-                                         orbit_info=info, update_files=True)
-        # force index to None
-        self.testInst.orbits.orbit_index = None
-        self.testInst.load(2009, 1)
-        with pytest.raises(ValueError):
-            self.testInst.orbits.next()
-
-    def test_orbit_orbibt_w_bad_variable(self):
-        info = {'index': 'magnetic local time', 'kind': 'orbit'}
-        self.testInst = pysat.Instrument('pysat', 'testing',
-                                         clean_level='clean',
-                                         orbit_info=info, update_files=True)
-        self.testInst.load(2009, 1)
-        with pytest.raises(ValueError):
-            self.testInst.orbits.next()
-
-    def test_orbit_orbit_w_missing_orbit_index(self):
-        info = {'index': 'magnetic local time', 'kind': 'orbit'}
-        self.testInst = pysat.Instrument('pysat', 'testing',
-                                         clean_level='clean',
-                                         orbit_info=info, update_files=True)
-        # force index to None
-        self.testInst.orbits.orbit_index = None
-        self.testInst.load(2009, 1)
-        with pytest.raises(ValueError):
-            self.testInst.orbits.next()
+        assert out_str.find("Orbit Settings") >= 0
+        assert out_str.find("Orbit Lind: local time") < 0
 
 
 class TestSpecificUTOrbits():
-
     def setup(self):
-        """Runs before every method to create a clean testing setup."""
-        info = {'index': 'mlt'}
+        """Runs before every method to create a clean testing setup
+        """
         self.testInst = pysat.Instrument('pysat', 'testing',
                                          clean_level='clean',
-                                         orbit_info=info, update_files=True)
+                                         orbit_info={'index': 'mlt'},
+                                         update_files=True)
+        self.stime = dt.datetime(2009, 1, 1)
+        self.inc_min = 97
+        self.etime = None
 
     def teardown(self):
-        """Runs after every method to clean up previous testing."""
-        del self.testInst
+        """Runs after every method to clean up previous testing
+        """
+        del self.testInst, self.stime, self.inc_min, self.etime
 
-    def test_single_orbit_call_by_0_index(self):
-        self.testInst.load(2009, 1)
-        self.testInst.orbits[0]
-        ans = (self.testInst.index[0] == dt.datetime(2009, 1, 1))
-        ans2 = (self.testInst.index[-1] == dt.datetime(2009, 1, 1, 1, 36, 59))
-        assert ans & ans2
+    @pytest.mark.parametrize('orbit_inc', [(0), (1), (-1), (-2), (14)])
+    def test_single_orbit_call_by_index(self, orbit_inc):
+        """Test successful orbit call by index
+        """
+        # Load the data
+        self.testInst.load(date=self.stime)
+        self.testInst.orbits[orbit_inc]
 
-    def test_single_orbit_call_by_1_index(self):
-        self.testInst.load(2009, 1)
-        self.testInst.orbits[1]
-        ans = (self.testInst.index[0] == dt.datetime(2009, 1, 1, 1, 37))
-        ans2 = (self.testInst.index[-1] == dt.datetime(2009, 1, 1, 3, 13, 59))
-        assert ans & ans2
+        # Increment the time
+        if orbit_inc >= 0:
+            self.stime += dt.timedelta(minutes=orbit_inc * self.inc_min)
+        else:
+            self.stime += dt.timedelta(minutes=self.inc_min
+                                       * (np.ceil(1440.0 / self.inc_min)
+                                          + orbit_inc))
+        self.etime = self.stime + dt.timedelta(seconds=(self.inc_min * 60 - 1))
 
-    def test_single_orbit_call_by_negative_1_index(self):
-        self.testInst.load(2008, 366)
-        self.testInst.orbits[-1]
-        ans = (self.testInst.index[0]
-               == (dt.datetime(2009, 1, 1)
-                   - relativedelta(hours=1, minutes=37)))
-        ans2 = (self.testInst.index[-1]
-                == (dt.datetime(2009, 1, 1)
-                    - relativedelta(seconds=1)))
-        assert ans & ans2
+        # Test the time
+        assert (self.testInst.index[0] == self.stime)
+        assert (self.testInst.index[-1] == self.etime)
 
-    def test_single_orbit_call_by_last_index(self):
-        self.testInst.load(2008, 366)
-        self.testInst.orbits[14]
-        assert (self.testInst.index[0]
-                == (dt.datetime(2009, 1, 1)
-                    - relativedelta(hours=1, minutes=37)))
-        assert (self.testInst.index[-1]
-                == (dt.datetime(2009, 1, 1)
-                    - relativedelta(seconds=1)))
-
-    def test_single_orbit_call_too_many(self):
-        self.testInst.load(2008, 366)
-        with pytest.raises(Exception):
-            self.testInst.orbits[17]
-
-    def test_single_orbit_no_input(self):
-        self.testInst.load(2008, 366)
-        with pytest.raises(TypeError):
-            self.testInst.orbits[None]
+    @pytest.mark.parametrize("orbit_ind,raise_err", [(17, Exception),
+                                                     (None, TypeError)])
+    def test_single_orbit_call_bad_index(self, orbit_ind, raise_err):
+        """ Test orbit failure with bad index
+        """
+        self.testInst.load(date=self.stime)
+        with pytest.raises(raise_err):
+            self.testInst.orbits[orbit_ind]
 
     def test_oribt_number_via_current_multiple_orbit_calls_in_day(self):
-        self.testInst.load(2009, 1)
-        self.testInst.bounds = (dt.datetime(2009, 1, 1), None)
+        """ Test orbit number with mulitple orbits calls in a day
+        """
+        self.testInst.load(date=self.stime)
+        self.testInst.bounds = (self.stime, None)
         true_vals = np.arange(15)
         true_vals[-1] = 0
         test_vals = []
@@ -179,29 +152,32 @@ class TestSpecificUTOrbits():
         assert np.all(test_vals == true_vals)
 
     def test_all_single_orbit_calls_in_day(self):
-        self.testInst.load(2009, 1)
-        ans = []
-        ans2 = []
-        self.testInst.bounds = (dt.datetime(2009, 1, 1), None)
+        """ Test all single orbit calls in a day
+        """
+        self.testInst.load(date=self.stime)
+        self.testInst.bounds = (self.stime, None)
         for i, inst in enumerate(self.testInst.orbits):
             if i > 14:
                 break
-            ans.append(self.testInst.index[0]
-                       == (dt.datetime(2009, 1, 1)
-                           + i * relativedelta(hours=1, minutes=37)))
-            ans2.append(self.testInst.index[-1]
-                        == (dt.datetime(2009, 1, 1)
-                            + (i + 1) * relativedelta(hours=1, minutes=37)
-                            - relativedelta(seconds=1)))
 
-        assert np.all(ans) & np.all(ans2)
+            # Test the start index
+            self.etime = self.stime + i * relativedelta(minutes=self.inc_min)
+            assert self.testInst.index[0] == self.etime
+
+            # Test the end index
+            self.etime += relativedelta(seconds=((self.inc_min * 60) - 1))
+            assert self.testInst.index[-1] == self.etime
 
     def test_orbit_next_call_no_loaded_data(self):
+        """ Test orbit next call without loading data
+        """
         self.testInst.orbits.next()
         assert (self.testInst.index[0] == dt.datetime(2008, 1, 1))
         assert (self.testInst.index[-1] == dt.datetime(2008, 1, 1, 0, 38, 59))
 
     def test_orbit_prev_call_no_loaded_data(self):
+        """ Test orbit previous call without loading data
+        """
         self.testInst.orbits.prev()
         # this isn't a full orbit
         assert (self.testInst.index[-1]
@@ -209,37 +185,43 @@ class TestSpecificUTOrbits():
         assert (self.testInst.index[0] == dt.datetime(2010, 12, 31, 23, 49))
 
     def test_single_orbit_call_orbit_starts_0_UT_using_next(self):
-        self.testInst.load(2009, 1)
+        """ Test orbit next call with data
+        """
+        self.testInst.load(date=self.stime)
         self.testInst.orbits.next()
-        assert (self.testInst.index[0] == dt.datetime(2009, 1, 1))
-        assert (self.testInst.index[-1] == dt.datetime(2009, 1, 1, 1, 36, 59))
+        self.etime = self.stime + dt.timedelta(seconds=(self.inc_min * 60 - 1))
+        assert (self.testInst.index[0] == self.stime)
+        assert (self.testInst.index[-1] == self.etime)
 
     def test_single_orbit_call_orbit_starts_0_UT_using_prev(self):
-        self.testInst.load(2009, 1)
+        """ Test orbit prev call with data
+        """
+        self.testInst.load(date=self.stime)
         self.testInst.orbits.prev()
-        assert (self.testInst.index[0]
-                == (dt.datetime(2009, 1, 1)
-                    + 14 * relativedelta(hours=1, minutes=37)))
-        assert (self.testInst.index[-1]
-                == (dt.datetime(2009, 1, 1)
-                    + 15 * relativedelta(hours=1, minutes=37)
-                    - relativedelta(seconds=1)))
+        self.stime += 14 * relativedelta(minutes=self.inc_min)
+        self.etime = self.stime + dt.timedelta(seconds=((self.inc_min * 60)
+                                                        - 1))
+        assert self.testInst.index[0] == self.stime
+        assert self.testInst.index[-1] == self.etime
 
     def test_single_orbit_call_orbit_starts_off_0_UT_using_next(self):
-        from dateutil.relativedelta import relativedelta as relativedelta
-        self.testInst.load(2008, 366)
+        """ Test orbit next call with data for previous day
+        """
+        self.stime -= dt.timedelta(days=1)
+        self.testInst.load(date=self.stime)
         self.testInst.orbits.next()
         assert (self.testInst.index[0] == dt.datetime(2008, 12, 30, 23, 45))
         assert (self.testInst.index[-1]
                 == (dt.datetime(2008, 12, 30, 23, 45)
-                    + relativedelta(hours=1, minutes=36, seconds=59)))
+                    + relativedelta(seconds=(self.inc_min * 60 - 1))))
 
     def test_single_orbit_call_orbit_starts_off_0_UT_using_prev(self):
-        self.testInst.load(2008, 366)
+        self.stime -= dt.timedelta(days=1)
+        self.testInst.load(date=self.stime)
         self.testInst.orbits.prev()
         assert (self.testInst.index[0]
                 == (dt.datetime(2009, 1, 1)
-                    - relativedelta(hours=1, minutes=37)))
+                    - relativedelta(minutes=self.inc_min)))
         assert (self.testInst.index[-1]
                 == (dt.datetime(2009, 1, 1) - relativedelta(seconds=1)))
 
@@ -247,10 +229,10 @@ class TestSpecificUTOrbits():
 class TestGeneralOrbitsMLT():
     def setup(self):
         """Runs before every method to create a clean testing setup."""
-        info = {'index': 'mlt'}
         self.testInst = pysat.Instrument('pysat', 'testing',
                                          clean_level='clean',
-                                         orbit_info=info, update_files=True)
+                                         orbit_info={'index': 'mlt'},
+                                         update_files=True)
 
     def teardown(self):
         """Runs after every method to clean up previous testing."""
@@ -463,10 +445,10 @@ class TestGeneralOrbitsMLT():
 class TestGeneralOrbitsMLTxarray(TestGeneralOrbitsMLT):
     def setup(self):
         """Runs before every method to create a clean testing setup."""
-        info = {'index': 'mlt'}
         self.testInst = pysat.Instrument('pysat', 'testing_xarray',
                                          clean_level='clean',
-                                         orbit_info=info, update_files=True)
+                                         orbit_info={'index': 'mlt'},
+                                         update_files=True)
 
     def teardown(self):
         """Runs after every method to clean up previous testing."""
@@ -477,10 +459,11 @@ class TestGeneralOrbitsLong(TestGeneralOrbitsMLT):
 
     def setup(self):
         """Runs before every method to create a clean testing setup."""
-        info = {'index': 'longitude', 'kind': 'longitude'}
         self.testInst = pysat.Instrument('pysat', 'testing',
                                          clean_level='clean',
-                                         orbit_info=info, update_files=True)
+                                         orbit_info={'index': 'longitude',
+                                                     'kind': 'longitude'},
+                                         update_files=True)
 
     def teardown(self):
         """Runs after every method to clean up previous testing."""
@@ -491,10 +474,11 @@ class TestGeneralOrbitsLongxarray(TestGeneralOrbitsMLT):
 
     def setup(self):
         """Runs before every method to create a clean testing setup."""
-        info = {'index': 'longitude', 'kind': 'longitude'}
         self.testInst = pysat.Instrument('pysat', 'testing_xarray',
                                          clean_level='clean',
-                                         orbit_info=info, update_files=True)
+                                         orbit_info={'index': 'longitude',
+                                                     'kind': 'longitude'},
+                                         update_files=True)
 
     def teardown(self):
         """Runs after every method to clean up previous testing."""
@@ -505,10 +489,11 @@ class TestGeneralOrbitsOrbitNumber(TestGeneralOrbitsMLT):
 
     def setup(self):
         """Runs before every method to create a clean testing setup."""
-        info = {'index': 'orbit_num', 'kind': 'orbit'}
         self.testInst = pysat.Instrument('pysat', 'testing',
                                          clean_level='clean',
-                                         orbit_info=info, update_files=True)
+                                         orbit_info={'index': 'orbit_num',
+                                                     'kind': 'orbit'},
+                                         update_files=True)
 
     def teardown(self):
         """Runs after every method to clean up previous testing."""
@@ -519,10 +504,11 @@ class TestGeneralOrbitsOrbitNumberXarray(TestGeneralOrbitsMLT):
 
     def setup(self):
         """Runs before every method to create a clean testing setup."""
-        info = {'index': 'orbit_num', 'kind': 'orbit'}
         self.testInst = pysat.Instrument('pysat', 'testing_xarray',
                                          clean_level='clean',
-                                         orbit_info=info, update_files=True)
+                                         orbit_info={'index': 'orbit_num',
+                                                     'kind': 'orbit'},
+                                         update_files=True)
 
     def teardown(self):
         """Runs after every method to clean up previous testing."""
@@ -533,10 +519,11 @@ class TestGeneralOrbitsLatitude(TestGeneralOrbitsMLT):
 
     def setup(self):
         """Runs before every method to create a clean testing setup."""
-        info = {'index': 'latitude', 'kind': 'polar'}
         self.testInst = pysat.Instrument('pysat', 'testing',
                                          clean_level='clean',
-                                         orbit_info=info, update_files=True)
+                                         orbit_info={'index': 'latitude',
+                                                     'kind': 'polar'},
+                                         update_files=True)
 
     def teardown(self):
         """Runs after every method to clean up previous testing."""
@@ -547,10 +534,11 @@ class TestGeneralOrbitsLatitudeXarray(TestGeneralOrbitsMLT):
 
     def setup(self):
         """Runs before every method to create a clean testing setup."""
-        info = {'index': 'latitude', 'kind': 'polar'}
         self.testInst = pysat.Instrument('pysat', 'testing_xarray',
                                          clean_level='clean',
-                                         orbit_info=info, update_files=True)
+                                         orbit_info={'index': 'latitude',
+                                                     'kind': 'polar'},
+                                         update_files=True)
 
     def teardown(self):
         """Runs after every method to clean up previous testing."""
@@ -583,10 +571,10 @@ def filter_data2(inst, times=None):
 class TestOrbitsGappyData(TestGeneralOrbitsMLT):
     def setup(self):
         """Runs before every method to create a clean testing setup."""
-        info = {'index': 'mlt'}
         self.testInst = pysat.Instrument('pysat', 'testing',
                                          clean_level='clean',
-                                         orbit_info=info, update_files=True)
+                                         orbit_info={'index': 'mlt'},
+                                         update_files=True)
         self.testInst.custom.attach(filter_data, 'modify')
 
     def teardown(self):
@@ -597,28 +585,27 @@ class TestOrbitsGappyData(TestGeneralOrbitsMLT):
 class TestOrbitsGappyDataXarray(TestGeneralOrbitsMLT):
     def setup(self):
         """Runs before every method to create a clean testing setup."""
-        info = {'index': 'mlt'}
         self.testInst = pysat.Instrument('pysat', 'testing_xarray',
                                          clean_level='clean',
-                                         orbit_info=info, update_files=True)
+                                         orbit_info={'index': 'mlt'},
+                                         update_files=True)
         self.testInst.custom.attach(filter_data, 'modify')
 
     def teardown(self):
         """Runs after every method to clean up previous testing."""
+        del self.testInst
 
 
 class TestOrbitsGappyData2(TestGeneralOrbitsMLT):
     def setup(self):
         """Runs before every method to create a clean testing setup."""
-        info = {'index': 'mlt'}
         self.testInst = pysat.Instrument('pysat', 'testing',
                                          clean_level='clean',
-                                         orbit_info=info)
+                                         orbit_info={'index': 'mlt'})
         times = [[dt.datetime(2008, 12, 31, 4),
                   dt.datetime(2008, 12, 31, 5, 37)],
                  [dt.datetime(2009, 1, 1),
-                  dt.datetime(2009, 1, 1, 1, 37)]
-                 ]
+                  dt.datetime(2009, 1, 1, 1, 37)]]
         for seconds in np.arange(38):
             day = (dt.datetime(2009, 1, 2)
                    + pds.DateOffset(days=int(seconds)))
@@ -637,15 +624,13 @@ class TestOrbitsGappyData2(TestGeneralOrbitsMLT):
 class TestOrbitsGappyData2Xarray(TestGeneralOrbitsMLT):
     def setup(self):
         """Runs before every method to create a clean testing setup."""
-        info = {'index': 'mlt'}
         self.testInst = pysat.Instrument('pysat', 'testing_xarray',
                                          clean_level='clean',
-                                         orbit_info=info)
+                                         orbit_info={'index': 'mlt'})
         times = [[dt.datetime(2008, 12, 31, 4),
                   dt.datetime(2008, 12, 31, 5, 37)],
                  [dt.datetime(2009, 1, 1),
-                  dt.datetime(2009, 1, 1, 1, 37)]
-                 ]
+                  dt.datetime(2009, 1, 1, 1, 37)]]
         for seconds in np.arange(38):
             day = (dt.datetime(2009, 1, 2)
                    + pds.DateOffset(days=int(seconds)))
@@ -664,10 +649,10 @@ class TestOrbitsGappyData2Xarray(TestGeneralOrbitsMLT):
 class TestOrbitsGappyLongData(TestGeneralOrbitsMLT):
     def setup(self):
         """Runs before every method to create a clean testing setup."""
-        info = {'index': 'longitude', 'kind': 'longitude'}
         self.testInst = pysat.Instrument('pysat', 'testing',
                                          clean_level='clean',
-                                         orbit_info=info)
+                                         orbit_info={'index': 'longitude',
+                                                     'kind': 'longitude'})
         self.testInst.custom.attach(filter_data, 'modify')
 
     def teardown(self):
@@ -678,10 +663,10 @@ class TestOrbitsGappyLongData(TestGeneralOrbitsMLT):
 class TestOrbitsGappyLongDataXarray(TestGeneralOrbitsMLT):
     def setup(self):
         """Runs before every method to create a clean testing setup."""
-        info = {'index': 'longitude', 'kind': 'longitude'}
         self.testInst = pysat.Instrument('pysat', 'testing_xarray',
                                          clean_level='clean',
-                                         orbit_info=info)
+                                         orbit_info={'index': 'longitude',
+                                                     'kind': 'longitude'})
         self.testInst.custom.attach(filter_data, 'modify')
 
     def teardown(self):
@@ -692,10 +677,10 @@ class TestOrbitsGappyLongDataXarray(TestGeneralOrbitsMLT):
 class TestOrbitsGappyOrbitNumData(TestGeneralOrbitsMLT):
     def setup(self):
         """Runs before every method to create a clean testing setup."""
-        info = {'index': 'orbit_num', 'kind': 'orbit'}
         self.testInst = pysat.Instrument('pysat', 'testing',
                                          clean_level='clean',
-                                         orbit_info=info)
+                                         orbit_info={'index': 'orbit_num',
+                                                     'kind': 'orbit'})
         self.testInst.custom.attach(filter_data, 'modify')
 
     def teardown(self):
@@ -706,10 +691,10 @@ class TestOrbitsGappyOrbitNumData(TestGeneralOrbitsMLT):
 class TestOrbitsGappyOrbitNumDataXarray(TestGeneralOrbitsMLT):
     def setup(self):
         """Runs before every method to create a clean testing setup."""
-        info = {'index': 'orbit_num', 'kind': 'orbit'}
         self.testInst = pysat.Instrument('pysat', 'testing_xarray',
                                          clean_level='clean',
-                                         orbit_info=info)
+                                         orbit_info={'index': 'orbit_num',
+                                                     'kind': 'orbit'})
         self.testInst.custom.attach(filter_data, 'modify')
 
     def teardown(self):
@@ -720,10 +705,10 @@ class TestOrbitsGappyOrbitNumDataXarray(TestGeneralOrbitsMLT):
 class TestOrbitsGappyOrbitLatData(TestGeneralOrbitsMLT):
     def setup(self):
         """Runs before every method to create a clean testing setup."""
-        info = {'index': 'latitude', 'kind': 'polar'}
         self.testInst = pysat.Instrument('pysat', 'testing',
                                          clean_level='clean',
-                                         orbit_info=info)
+                                         orbit_info={'index': 'latitude',
+                                                     'kind': 'polar'})
         self.testInst.custom.attach(filter_data, 'modify')
 
     def teardown(self):
@@ -734,10 +719,10 @@ class TestOrbitsGappyOrbitLatData(TestGeneralOrbitsMLT):
 class TestOrbitsGappyOrbitLatDataXarray(TestGeneralOrbitsMLT):
     def setup(self):
         """Runs before every method to create a clean testing setup."""
-        info = {'index': 'latitude', 'kind': 'polar'}
         self.testInst = pysat.Instrument('pysat', 'testing_xarray',
                                          clean_level='clean',
-                                         orbit_info=info)
+                                         orbit_info={'index': 'latitude',
+                                                     'kind': 'polar'})
         self.testInst.custom.attach(filter_data, 'modify')
 
     def teardown(self):